requirements.txt

# Minimum required Python version: 3.10+
# Dependencies discovered by scanning project imports
streamlit>=1.20.0
groq>=0.1.0
python-dotenv>=1.0.0
GitPython>=3.1.0
fpdf2>=2.5.0
Markdown>=3.4.0

